Load palette for particles.

+++ GameVM/code/cgame/cg_local.h
@@ -53,7 +53,7 @@ typedef struct particle_s
 {
 // driver-usable fields
 	vec3_t		org;
-	float		color;
+	byte		color;
 // drivers never touch the following fields
 	struct particle_s	*next;
 	vec3_t		vel;
